--- src/components/kanban2/CreateTopic.js
@@ -1,60 +1,0 @@
-import React from 'react';
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import { makeStyles } from '@material-ui/core';
-
-const useStyles = makeStyles(theme => ({
-  container: {
-    position: 'absolute',
-    top: theme.spacing(2),
-    left: theme.spacing(2),
-    display: 'flex',
-    justifyContent:'center'
-  }
-}));
-
-export default function CreateTopic() {
-  const { container } = useStyles();
-  const [open, setOpen] = React.useState(false);
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <div class={container}>
-      <Button variant="outlined" color="secondary" onClick={handleClickOpen}>
-        Add Topic
-      </Button>
-      <Dialog open={open} onClose={handleClose} aria-labelledby="form-dialog-title">
-        <DialogTitle id="form-dialog-title">Add a Topic</DialogTitle>
-        <DialogContent>
-          <TextField
-            autoFocus
-            margin="dense"
-            id="name"
-            label="Topic Title"
-            type="Topic"
-            fullWidth
-          />
-        </DialogContent>
-        <DialogActions style={{display: 'flex', justifyContent:'space-between'}}>
-          <Button onClick={handleClose} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={handleClose} color="primary">
-            Add
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
-  );
-}
